+++ src/test/java/lotto/LottoTest.java
@@ -1,4 +1,4 @@
-package lotto.model.domain;
+package lotto;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
